Refactor: revert로 생성된 passport 모듈 제거
