+++ src/domains/auth/consts/platform.const.ts
