+++ src/domains/user/entities/user.entity.ts
